requirements.txt
langchain
openai
py-solc-x
slither-analyzer
rich
markdown
